Rename files and organize login page.
