+++ together-wherever/src/fetcher/createNewTrip.ts
