+++ together-wherever/src/fetcher/getTripDetails.ts
