+++ together-wherever/src/hooks/useCreateNewTrip.ts
@@ -6,7 +6,7 @@ import { useRouter } from "next/navigation";
 import { jwtDecode } from 'jwt-decode';
 
 import { CreateNewTripBodyInterface } from "@/utils/types";
-import { fetchingUsersData, createNewTrip } from "@/fetcher/create-new-trip";
+import { fetchingUsersData, createNewTrip } from "@/fetcher/createNewTrip";
 import { fetchPlaceDetails } from "@/fetcher/getPlaceDetails";
 
 export const useCreateNewTrips = () => {
